+++ aliyun-python-sdk-core/alibabacloud/endpoint/location_endpoint_caller.py
@@ -14,7 +14,6 @@
 
 from alibabacloud.client import AlibabaCloudClient
 from alibabacloud.request import APIRequest
-DEFAULT_LOCATION_SERVICE_ENDPOINT = "location-readonly.aliyuncs.com"
 
 
 class MiniLocationClient(AlibabaCloudClient):
@@ -26,14 +25,14 @@ class MiniLocationClient(AlibabaCloudClient):
         self.location_service_code = None
         self.location_endpoint_type = "openAPI"
 
-    def describe_endpoint(self, region_id=None, endpoint_type=None, location_service_code=None):
+    def describe_endpoint(self, region_id=None, endpoint_type=None, location_service_code=None, location_endpoint=None):
         api_request = APIRequest('DescribeEndpoints', 'GET', 'https', 'RPC')
         api_request._params = {
             "Id": region_id,
             "Type": endpoint_type,
             "ServiceCode": location_service_code,
         }
-        api_request.endpoint = DEFAULT_LOCATION_SERVICE_ENDPOINT
+        api_request.endpoint = location_endpoint
         return self._handle_request(api_request)
 
 
@@ -42,5 +41,5 @@ class DescribeEndpointCaller:
     def __init__(self, client_config, credentials_provider):
         self._client = MiniLocationClient(client_config, credentials_provider)
 
-    def fetch(self, region_id, endpoint_type, location_service_code):
-        return self._client.describe_endpoint(region_id, endpoint_type, location_service_code)
+    def fetch(self, region_id, endpoint_type, location_service_code, location_endpoint):
+        return self._client.describe_endpoint(region_id, endpoint_type, location_service_code, location_endpoint)
