--- aliyun-python-sdk-core/aliyunsdkcore/endpoint/location/DescribeEndpointsRequest.py
@@ -1,42 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-#
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-
-from aliyunsdkcore.request import RpcRequest
-class DescribeEndpointsRequest(RpcRequest):
-
-	def __init__(self):
-		RpcRequest.__init__(self, 'Location', '2015-06-12', 'DescribeEndpoints')
-
-	def get_ServiceCode(self):
-		return self.get_query_params().get('ServiceCode')
-
-	def set_ServiceCode(self,ServiceCode):
-		self.add_query_param('ServiceCode',ServiceCode)
-
-	def get_Id(self):
-		return self.get_query_params().get('Id')
-
-	def set_Id(self,Id):
-		self.add_query_param('Id',Id)
-
-	def get_Type(self):
-		return self.get_query_params().get('Type')
-
-	def set_Type(self,Type):
-		self.add_query_param('Type',Type)
