--- aliyun-python-sdk-core/aliyunsdkcore/profile/__init__.py
@@ -1,1 +1,0 @@
-__author__ = 'Alibaba Cloud'
